requirements.txt
filetype==1.2.0
Pillow==9.4.0
